# Add Electron support
It is now possible to run the Angular App as a desktop app with Electron!

To run the app using Electron, run the following npm commands:
- npm install
- npm run-script electron

This command will build the Angular app using Angular CLI, then the output will be copied in the electron\app folder. Finally (after some additional tasks), the app will run as a desktop app.

Some improvements are still needed in order to improve the process...